src/llm/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9

numpy==1.26.4
pandas==2.2.1
scikit-learn==1.2.2
scipy==1.10.1

beautifulsoup4==4.12.3

transformers==4.29.2
datasets==2.18.0

openai==1.16.2
langchain==0.1.14
langchain-openai==0.1.1
tiktoken==0.6.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
langchainhub==0.1.15


sentencepiece==0.1.99
nltk==3.8.1
spacy==3.5.3
prophet==1.1.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.0

tavily-python==0.3.3 # https://python.langchain.com/docs/integrations/retrievers/tavily/ TAVILY_API_KEY